Move tokens to a separate page and add to the sidebar
As a customer, I’m not likely to visit the `Users` page to create a
personal access token.

Let’s make it easier to find by moving it to its own page and adding it
to the sidebar.
