+++ views/project/dashboard.erb
@@ -78,7 +78,7 @@ cards = [
     "hero-key",
     "Create Access Token",
     "Create as many personal access tokens as you want and manage your resources programmatically.",
-    "#{@project_data[:path]}/user/token",
+    "#{@project_data[:path]}/token",
     "Project:token"
   ],
   [
