+++ views/project/token.erb
@@ -1,5 +1,4 @@
 <% @page_title = "#{@project_data[:name]} - Personal Access Tokens" %>
-<%== render("project/user-tabbar") %>
 
 <%== render(
   "components/page_header",
@@ -7,8 +6,8 @@
     title: "Personal Access Tokens",
     breadcrumbs: [%w[Projects /project], [@project_data[:name], @project_data[:path]], ["Personal Access Tokens", "#"]],
     right_items: [<<-CREATE]
-      <form id="create-pat" action="#{@project_data[:path]}/user/token" role="form" method="POST">
-        #{csrf_tag("#{@project_data[:path]}/user/token")}
+      <form id="create-pat" action="#{@project_data[:path]}/token" role="form" method="POST">
+        #{csrf_tag("#{@project_data[:path]}/token")}
         #{render("components/form/submit_button", locals: { text: "Create Token" })}
       </form>
     CREATE
@@ -39,7 +38,7 @@
             <tr id="token-<%= token.ubid %>">
               <td class="whitespace-nowrap py-4 pl-4 pr-3 text-sm font-medium text-gray-900 sm:pl-6" scope="row">
                 <a
-                  href="<%= @project_data[:path] %>/user/token/<%= token.ubid %>/access-control"
+                  href="<%= @project_data[:path] %>/token/<%= token.ubid %>/access-control"
                   class="text-orange-600 hover:text-orange-700"
                 ><%= token.ubid %></a>
               </td>
@@ -51,7 +50,7 @@
                   "components/delete_button",
                   locals: {
                     text: "Remove",
-                    url: "#{@project_data[:path]}/user/token/#{token.ubid}",
+                    url: "#{@project_data[:path]}/token/#{token.ubid}",
                     confirmation: "delete token"
                   }
                 ) %>
